Remove unnecessary reciter from audio response (#75)
The audio structure in the json was repeated per ayah - this is not
necessary (as a matter of fact, it's not needed at all, since the
reciter id is already known before making this call). This saves a
database query and a very small amount of transfer (4k uncompressed,
much less compressed (~100 bytes)).