Don't return font numbers in search api (#76)
When we do a search, we also search the fonts - we use the information from
these to properly highlight the results. While this is all good, we should
not be returning these results back from the api. This patch fixes that by
returning a nil from that block, and then compacting the match array in
order to remove the nil element before returning. Fixes #63.